+++ hata/discord/channel/channel_guild_directory.py
@@ -198,7 +198,7 @@ class ChannelDirectory(ChannelGuildMainBase, ChannelTextBase):
     def precreate(cls, channel_id, **kwargs):
         """
         Precreates the channel by creating a partial one with the given parameters. When the channel is loaded
-        the precrated channel will be picked up. If an already existing channel would be precreated, returns that
+        the precreated channel will be picked up. If an already existing channel would be precreated, returns that
         instead and updates that only, if that is a partial channel.
         
         Parameters
@@ -258,3 +258,9 @@ class ChannelDirectory(ChannelGuildMainBase, ChannelTextBase):
                 setattr(self, *item)
         
         return self
+
+
+    @copy_docs(ChannelGuildMainBase.to_data)
+    def to_data(self):
+        # same as `ChannelGuildMainBase`'s. Keep this function here for reference.
+        return ChannelGuildMainBase.to_data(self)