+++ hata/discord/channel/channel_guild_undefined.py
@@ -232,7 +232,7 @@ class ChannelGuildUndefined(ChannelGuildMainBase):
     def precreate(cls, channel_id, **kwargs):
         """
         Precreates the channel by creating a partial one with the given parameters. When the channel is loaded
-        the precrated channel will be picked up. If an already existing channel would be precreated, returns that
+        the precreated channel will be picked up. If an already existing channel would be precreated, returns that
         instead and updates that only, if that is a partial channel.
         
         Parameters
@@ -302,3 +302,11 @@ class ChannelGuildUndefined(ChannelGuildMainBase):
                 setattr(self, *item)
         
         return self
+    
+    
+    @copy_docs(ChannelGuildMainBase.to_data)
+    def to_data(self):
+        data = ChannelGuildMainBase.to_data(self)
+        data.update(self.__dict__)
+        
+        return data